// src/users/utilities/enum.ts
export enum USER_ROLE {
  candidate = 'candidate',
  mentor = 'mentor',
  hiring_manager = 'hiring_manager',
  recruiter = 'recruiter',
  admin = 'admin',
}

export enum GENDER {
  male = 'male',
  female = 'female',
  non_binary = 'non_binary',
  lgbtq = 'lgbtq',
  prefer_not_to_say = 'prefer_not_to_say',
}

export enum LoginMethodConstants {
  email = 'email',
  google = 'google',
}